+++ lib/wheel_chooser.dart
@@ -15,6 +15,7 @@ class WheelChooser extends StatefulWidget {
   final double listHeight;
   final double listWidth;
   final List<Widget> children;
+  final bool horizontal;
   static const double _defaultitemSize = 48.0;
 
   WheelChooser({
@@ -29,6 +30,7 @@ class WheelChooser extends StatefulWidget {
     this.perspective = 0.01,
     this.listWidth,
     this.listHeight,
+    this.horizontal = false,
   })  : assert(perspective <= 0.01),
         children = null;
 
@@ -43,6 +45,7 @@ class WheelChooser extends StatefulWidget {
     this.perspective = 0.01,
     this.listWidth,
     this.listHeight,
+    this.horizontal = false,
   })  : assert(perspective <= 0.01),
         assert(datas == null || datas.length == children.length),
         selectTextStyle = null,
@@ -62,18 +65,27 @@ class WheelChooser extends StatefulWidget {
     this.perspective = 0.01,
     this.listWidth,
     this.listHeight,
+    this.horizontal = false,
+    bool reverse = false,
   })  : assert(perspective <= 0.01),
         assert(minValue < maxValue),
         assert(initValue == null || initValue >= minValue),
         assert(initValue == null || maxValue >= initValue),
+        assert(step > 0),
         children = null,
-        datas = _createIntegerList(minValue, maxValue, step),
-        startPosition = initValue == null ? 0 : initValue - minValue;
+        datas = _createIntegerList(minValue, maxValue, step, reverse),
+        startPosition = initValue == null ? 0 : reverse ? (maxValue - initValue) ~/ step : (initValue - minValue) ~/ step;
 
-  static List<int> _createIntegerList(int minValue, int maxValue, int step) {
+  static List<int> _createIntegerList(int minValue, int maxValue, int step, bool reverse) {
     List<int> result = [];
-    for (int i = minValue; i <= maxValue; i += step) {
-      result.add(i);
+    if (reverse) {
+      for (int i = maxValue; i >= minValue; i -= step) {
+        result.add(i);
+      }
+    } else {
+      for (int i = minValue; i <= maxValue; i += step) {
+        result.add(i);
+      }
     }
     return result;
   }
@@ -107,32 +119,57 @@ class _WheelChooserState extends State<WheelChooser> {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-        height: widget.listHeight ?? double.infinity,
-        width: widget.listWidth ?? double.infinity,
-        child: ListWheelScrollView(
-          onSelectedItemChanged: _listener,
-          perspective: widget.perspective,
-          squeeze: widget.squeeze,
-          controller: fixedExtentScrollController,
-          physics: FixedExtentScrollPhysics(),
-          children: widget.children ?? _buildListItems(),
-          useMagnifier: true,
-          magnification: widget.magnification,
-          itemExtent: widget.itemSize,
-        ));
+    return RotatedBox(
+        quarterTurns: widget.horizontal ? 3 : 0,
+        child: Container(
+            height: widget.listHeight ?? double.infinity,
+            width: widget.listWidth ?? double.infinity,
+            child: ListWheelScrollView(
+              onSelectedItemChanged: _listener,
+              perspective: widget.perspective,
+              squeeze: widget.squeeze,
+              controller: fixedExtentScrollController,
+              physics: FixedExtentScrollPhysics(),
+              children: _convertListItems() ?? _buildListItems(),
+              useMagnifier: true,
+              magnification: widget.magnification,
+              itemExtent: widget.itemSize,
+            )));
   }
 
   List<Widget> _buildListItems() {
     List<Widget> result = [];
     for (int i = 0; i < widget.datas.length; i++) {
       result.add(
-        Text(
-          widget.datas[i].toString(),
-          style: i == currentPosition ? widget.selectTextStyle ?? null : widget.unSelectTextStyle ?? null,
+        RotatedBox(
+          quarterTurns: widget.horizontal ? 1 : 0,
+          child: Text(
+            widget.datas[i].toString(),
+            textAlign: TextAlign.center,
+            textScaleFactor: 1.5,
+            style: i == currentPosition ? widget.selectTextStyle ?? null : widget.unSelectTextStyle ?? null,
+          ),
         ),
       );
     }
     return result;
+  }
+
+  List<Widget> _convertListItems() {
+    if (widget.children == null) {
+      return null;
+    }
+    if (widget.horizontal) {
+      List<Widget> result = [];
+      for (int i = 0; i < widget.children.length; i++) {
+        result.add(RotatedBox(
+          quarterTurns: 1,
+          child: widget.children[i],
+        ));
+      }
+      return result;
+    } else {
+      return widget.children;
+    }
   }
 }