+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## [0.0.1]
 
 * NEW: Initial Release.
+
+## [0.1.0]
+
+* NEW: Support Horizontal.
+* NEW: Support Reverse Integer.
+* Fixed bug: about Integer step InitValue is not Current.